Bids form on to the right
Right sidebar:
    [bids form]
    [bids list]

Putting the bids on top would push the form to way below. What would
be important is to have the form visible so a new bidder can easily
make a bid. There for the form is best on top.

In the other hand, existing bids may bot be as important. But having
them on the right side could make the page weighted to the right.

Would be best to add the 'bids' to the left below the author info
by splitting the partial in to two.