Recent Exchanges - brought to present view
There were some style changes ( html tag wise ) in a prev commit. The
`_recent_exchanges.html.erb` file was slightly different from the other
2 - offers and requests.

Instead of editing the file that might be needed by another view, the
code was brought to the local view.

An atempt was made to populate the field using 'Excchanges' but the
values didn't seem to reflect in the location. The view may or may not
be looking well.